generalizing DFDistT for multiple DFBlocks.
Made sure that Matrix can be pretty big (size_t instead of int)

+++ src/util/matrix.h
@@ -131,7 +131,7 @@ class Matrix : public Matrix_base<double>, public std::enable_shared_from_this<M
 
     double orthog(const std::list<std::shared_ptr<const Matrix> > o);
 
-    void print(const std::string in = "", const int size = 10) const;
+    void print(const std::string in = "", const size_t size = 10) const;
 
 #ifdef HAVE_SCALAPACK
     // return a shared pointer to this ifndef HAVE_SCALAPACK
